[Silver V] Title: LCM, Time: 116 ms, Memory: 14180 KB -BaekjoonHub

+++ 백준/Silver/5347. LCM/LCM.java
@@ -16,7 +16,7 @@ public class Main {
             int A = Integer.parseInt(st.nextToken());
             int B = Integer.parseInt(st.nextToken());
 
-            System.out.println(((long)A * (long)B) / solve(A, B));
+            System.out.println(((long) A * B) / solve(A, B));
         }
         br.close();
     }
